+++ source/cpp_implementation/miscellaneous/interfaces/i_image_buffer.h
@@ -12,8 +12,8 @@ class IImageBuffer
 public:
 	virtual int width() const = 0;
 	virtual int height() const = 0;
-	virtual void set_pixel_value(size_t width_index, size_t height_index, const c_vector3 &pixel_color_value, size_t samples_per_pixel) = 0;
-	virtual c_vector3 get_rgb_pixel(size_t index) = 0;
+	virtual void set_pixel_value(int width_index, int height_index, const c_vector3 &pixel_color_value, int samples_per_pixel) = 0;
+	virtual c_vector3 get_rgb_pixel(int index) = 0;
 	virtual c_vector3 get_rgb_pixel(int width_index, int height_index) = 0;
 	virtual std::unique_ptr<std::vector<c_vector3>> buffer() = 0;
